requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Authlib==1.3.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.0
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5d968ni_yn/croot/certifi_1671487774636/work/certifi
cffi==1.15.1
charset-normalizer==3.1.0
chroma-hnswlib==0.7.3
chromadb==0.4.14
ci-info==0.3.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.1.3
configobj==5.0.8
configparser==6.0.0
cryptography==42.0.3
dataclasses-json==0.5.7
datasets==2.14.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
dnspython==2.4.2
dspy-ai==2.1.9
etelemetry==0.3.1
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.99.1
fastjsonschema==2.16.3
filelock==3.12.4
fitz==0.0.1.dev2
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
flatbuffers==23.5.26
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
gptcache==0.1.11
grpcio==1.60.1
grpcio-health-checking==1.60.1
grpcio-tools==1.60.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.3
httplib2==0.22.0
httptools==0.6.0
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.4.1
importlib-resources==6.1.0
ipdb==0.13.13
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
langchain==0.0.314
langchainplus-sdk==0.0.9
langsmith==0.0.43
llama-index==0.6.25.post1
loguru==0.7.2
looseversion==1.3.0
lxml==4.9.3
Mako==1.3.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mistune==2.0.5
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.2.1
newsapi-python==0.2.7
nibabel==5.2.0
nipype==1.8.6
notebook==6.5.4
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.24.2
onnxruntime==1.16.1
openai==0.28.1
openapi-schema-pydantic==1.2.4
optuna==3.4.0
overrides==7.4.0
packaging==23.1
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
pinecone-client==2.2.4
platformdirs==3.2.0
posthog==3.0.2
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.24.4
prov==2.0.0
psutil==5.9.4
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydot==2.0.0
Pygments==2.15.0
pymongo==4.5.0
PyMuPDF==1.23.16
PyMuPDFb==1.23.9
pyparsing==3.1.1
pypdf==3.9.1
PyPDF2==3.0.1
PyPika==0.48.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
pyxnat==1.6
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
rdflib==7.0.0
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
safetensors==0.4.2
scikit-learn==1.3.1
scipy==1.11.1
Send2Trash==1.8.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.16
sseclient-py==1.8.0
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.2.0
tiktoken==0.3.3
tinycss2==1.2.1
together==0.2.11
tokenizers==0.15.1
tomli==2.0.1
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
traits==6.3.2
transformers==4.37.2
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.8.0
uri-template==1.2.0
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.18.0
validators==0.22.0
watchfiles==0.21.0
wcwidth==0.2.6
weaviate-client==4.4.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.3.6
widgetsnbextension==4.0.7
wikipedia==1.4.0
xxhash==3.4.1
yarl==1.8.2
zipp==3.15.0
